functions. needs visual adjusting
