+++ App.js
@@ -1,122 +1,106 @@
-import { StatusBar } from "expo-status-bar";
-import Checkbox from 'expo-checkbox';
-import React, { useState } from 'react';
-import { Button, SafeAreaView, StyleSheet, Text, TextInput, View, FlatList, ScrollView } from "react-native";
+import React, { useState } from "react";
+import {
+    View,
+    TextInput,
+    Button,
+    FlatList,
+    TouchableOpacity,
+    Text,
+} from "react-native";
 
 export default function App() {
-    const [checked, setChecked] = useState(false);
-    const [item, setItem] = useState('');
-    const [notes, setNotes] = useState('');
     const [shoppingList, setShoppingList] = useState([]);
-    const [selected, setSelected] = useState([]);
+    const [item, setItem] = useState("");
+    const [notes, setNotes] = useState("");
 
     const handleAdd = () => {
-        setShoppingList([...shoppingList, { item, notes }]);
-        setItem('');
-        setNotes('');
+        setShoppingList([
+            ...shoppingList,
+            { id: Date.now(), item, notes, checked: false },
+        ]);
+        setItem("");
+        setNotes("");
     };
 
-    const handleDelete = () => {
-        const newList = [...shoppingList];
-        newList.filter((item) => item !== checked);
+    const handleDelete = (id) => {
+        const newList = shoppingList.filter((item) => item.id !== id);
         setShoppingList(newList);
     };
 
+    const handleCheck = (id) => {
+        const newList = shoppingList.map((item) =>
+            item.id === id ? { ...item, checked: !item.checked } : item
+        );
+        setShoppingList(newList);
+    };
 
+    const handleEdit = (id) => {
+        // Handle edit functionality here
+        console.log(`Edit item with id ${id}`);
+    };
+
+    const renderItem = ({ item }) => (
+        <TouchableOpacity
+            style={{
+                flexDirection: "row",
+                alignItems: "center",
+                justifyContent: "space-between",
+                padding: 10,
+            }}
+            onPress={() => handleCheck(item.id)}
+        >
+            <Text
+                style={{
+                    textDecorationLine: item.checked ? "line-through" : "none",
+                }}
+            >
+                {item.item}
+            </Text>
+            <View style={{ flexDirection: "row" }}>
+                <TouchableOpacity onPress={() => handleEdit(item.id)}>
+                    <Text style={{ color: "blue", marginRight: 10 }}>Edit</Text>
+                </TouchableOpacity>
+                <TouchableOpacity onPress={() => handleDelete(item.id)}>
+                    <Text style={{ color: "red" }}>Delete</Text>
+                </TouchableOpacity>
+            </View>
+        </TouchableOpacity>
+    );
 
     return (
-        <SafeAreaView style={styles.container}>
-            {/* <ScrollView> */}
-                <View style={styles.form} >
-                    <Text>Item: </Text>
-                    <TextInput 
-                        style={styles.input}
-                        onChangeText={text => setItem(text)}
-                        value={item}
-                    />
-                    <Text>Notes: </Text>
-                    <TextInput 
-                        style={styles.input}
-                        onChangeText={text => setNotes(text)}
-                        value={notes}
-                    />
-                    <Button
-                        title="Add"
-                        onPress={handleAdd}
-                        color='purple'
-                    />
-                </View>
-                <View>
-                    <Text style={styles.title}>Shopping List</Text>
-                    {shoppingList.map((item, index) => (
-                        <FlatList
-                            key={(index) => index.toString()}
-                            data={shoppingList}
-                            renderItem={({ item }) => (
-                                <View style={styles.list}>
-                                    <Checkbox
-                                        value={checked}
-                                        onValueChange={() => {
-                                            setChecked(!checked);
-                                            setSelected(selected);
-                                        }}
-                                        style={styles.checkbox}
-                                    />
-                                    <Text>{item.item}</Text>
-                                    <Text>{item.notes}</Text>
-                                </View>
-                            )}
-                        />
-                    ))}
-                    <View style={styles.delete} >
-                        <Button
-                            title="Delete Checked Item"
-                            onPress={() => handleDelete()}
-                            color='red'
-                        />
-                    </View>
-                </View>
-                <StatusBar style="auto" />
-            {/* </ScrollView> */}
-        </SafeAreaView>
+        <View style={{ flex: 1, padding: 20, paddingTop: 55}}>
+            <View style={{ flexDirection: "row", marginBottom: 10 }}>
+                <TextInput
+                    style={{
+                        flex: 1,
+                        marginRight: 10,
+                        padding: 6,
+                        borderWidth: 1,
+                        borderRadius: 4,
+                    }}
+                    placeholder="Item"
+                    value={item}
+                    onChangeText={(text) => setItem(text)}
+                />
+                <TextInput
+                    style={{
+                        flex: 1,
+                        marginRight: 10,
+                        padding: 6,
+                        borderWidth: 1,
+                        borderRadius: 4,
+                    }}
+                    placeholder="Notes"
+                    value={notes}
+                    onChangeText={(text) => setNotes(text)}
+                />
+                <Button title="Add" onPress={handleAdd} />
+            </View>
+            <FlatList
+                data={shoppingList}
+                renderItem={renderItem}
+                keyExtractor={(item) => item.id.toString()}
+            />
+        </View>
     );
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: "#fff",
-        marginTop: 35,
-        marginLeft: 15,
-        alignContent: 'center',
-        marginRight: 15,
-
-    },
-    input: {
-        height: 40,
-        width: '50%',
-        margin: 12,
-        borderWidth: 1,
-        padding: 10,
-    },
-    form: {
-        flexDirection: "column",
-        justifyContent: "space-between",
-    },
-    list: {
-        marginTop: 20,
-    },
-    title: {
-        justifyContent: 'center',
-        fontSize: 24,
-        fontStyle: 'italic',
-        color: 'purple',
-        maxWidth: '50%',
-        marginLeft: 20,
-        marginTop: 20,
-    },
-    delete: {
-        marginTop: 20,
-        alignContent: 'bottom'
-    },
-});
